Move Medications into a combined Notes & Meds tab, tweak Feedings card
- Moved the Medications section from the Overview tab into the third
  tab, renamed "Notes & Meds" with a new combined clipboard icon
  (was just "Notes"/sticky-note).
- Feedings quick-stat card: replaced the redundant "9 feedings today"
  sub-line with "9 today" plus a new "Last: 1h 35m ago" line so it's
  clear at a glance how long since the last feeding.

Verified against the real Baby Buddy instance.
